requirements.txt
# AWS SDK
boto3>=1.28.0,<2.0.0
aws-lambda-powertools>=2.30.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# HTTP and API
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Audio Processing
pydub>=0.25.1
librosa>=0.10.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0 